metric sparse util

+++ profiling_tools/clibs/calcBin.c
@@ -21,12 +21,12 @@ int binary_srch(long long *arr, int l, int r, long long x) {
 	return r;
 }
 
-void calcHistogram(int *histogram_counter, int histogram_size,
+void calcHistogram(long long *histogram_counter, int histogram_size,
 		long long *histogram_index,
 		double *histogram_util, 
 		long long *critical_points, int critical_points_size,
 		long long *location_index, int location_size,
-		int *location_counter,
+		long long *location_counter,
 		double *location_util) {
 	int i;
 	int nextRecordIndex = 0;
